+++ src/shared/components/query/old/styles.module.scss
@@ -115,7 +115,7 @@
   }
 
   .searchTextInput {
-    minWidth: 200px;
+    min-width: 200px;
   }
 }
